# docs/swagger.yaml
...
basePath: /api/v1
definitions:
  github_com_MeshCore-Beacon_beacon-server_internal_api.AdvertObservation:
    properties:
      heardAt:
        description: epoch ms
        type: integer
      hopCount:
        type: integer
      iata:
        type: string
      id:
        description: observation ID, use as cursor for pagination
        type: integer
      nodeName:
        type: string
      nodePublicKey:
        description: hex-encoded
        type: string
      packetHash:
        description: hex-encoded
        type: string
      payloadType:
        type: integer
      payloadTypeName:
        type: string
      rssi:
        type: integer
      snr:
        type: number
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.Observer:
    properties:
      batteryLevel:
        description: volts, nil if mains powered
        type: number
      brokers:
        description: broker names this observer has been seen on
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverBroker'
        type: array
      displayName:
        description: friendly name from /status messages
        type: string
      firmwareBuild:
        type: string
      firmwareVersion:
        type: string
      firstSeen:
        description: epoch ms
        type: integer
      hardwareModel:
        type: string
      iata:
        description: most recently heard IATA
        type: string
      id:
        type: string
      lastSeen:
        description: epoch ms
        type: integer
      lastStatusAt:
        description: epoch ms
        type: integer
      observationCount:
        type: integer
      observerType:
        description: e.g. "meshcoretomqtt", "meshcoreha"
        type: string
      publicKey:
        description: hex-encoded public key
        type: string
      radio:
        description: 'friendly radio param string: freqMhz,BwKhz,SF'
        type: string
      radioBwKhz:
        description: bandwidth in kHz
        type: number
      radioCr:
        description: coding rate denominator
        type: integer
      radioFreqMhz:
        description: MHz e.g. 910.525
        type: number
      radioSf:
        description: LoRa spreading factor
        type: integer
      scopes:
        description: list of observer forwarded scopes matched to config
        items:
          type: string
        type: array
      softwareVersion:
        type: string
      status:
        description: '"online" or "offline" derived from last_status_at'
        type: string
      statusMetadata:
        description: raw /status JSON payload
      uptimeSeconds:
        type: integer
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverBroker:
    properties:
      lastPacketAt:
        description: epoch ms, last packet received via this broker; 0 if none
        type: integer
      lastSeenAt:
        description: epoch ms, last time observer was seen on this broker
        type: integer
      name:
        description: broker name e.g. "mqtt1"
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverSummary:
    properties:
      displayName:
        description: friendly name from /status messages
        type: string
      iata:
        description: most recently heard IATA
        type: string
      id:
        type: string
      observerType:
        description: e.g. "meshcoretomqtt", "meshcoreha"
        type: string
      radio:
        description: 'friendly radio param string: freqMhz,BwKhz,SF'
        type: string
      scopes:
        description: list of observer forwarded scopes matched to config
        items:
          type: string
        type: array
      status:
        description: '"online" or "offline" derived from last_status_at'
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetry:
    properties:
      interval:
        type: string
      points:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetryPoint'
        type: array
      range:
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetryPoint:
    properties:
      airtimeRxPct:
        type: number
      airtimeTxPct:
        type: number
      batteryMv:
        type: integer
      noiseFloorDb:
        type: number
      queueLength:
        type: integer
      receiveErrors:
        type: integer
      t:
        description: epoch ms
        type: integer
      uptimeSeconds:
        type: integer
    type: object
...
  ? github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_AdvertObservation
  : properties:
      hasMore:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.AdvertObservation'
        type: array
      nextCursor:
        type: integer
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_ObserverSummary:
    properties:
      hasMore:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverSummary'
        type: array
      nextCursor:
        type: integer
    type: object
...
  internal_api_handlers.APIError:
    properties:
      code:
        description: e.g. "not_found", "bad_request"
        type: string
      message:
        description: e.g. "channel not found"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: MeshCore Beacon
    url: https://github.com/MeshCore-Beacon/beacon-server
  description: MeshCore network observation backend. Ingests LoRa packets from MQTT
    brokers, stores in PostgreSQL, and streams live events via WebSocket.
  license:
    name: MIT
  termsOfService: https://github.com/MeshCore-Beacon/beacon-server
  title: MeshCore Beacon API
  version: "1.0"
paths:
...
  /observers:
    get:
      parameters:
      - description: Filter by single IATA code (case-insensitive)
        in: query
        name: iata
        type: string
      - description: Filter by multiple IATA codes, comma-separated e.g. YVR,YYJ
        in: query
        name: iatas
        type: string
      - description: Filter by region ID, expands to member IATAs
        in: query
        name: regionId
        type: integer
      - description: Filter by region slug, expands to member IATAs
        in: query
        name: region
        type: string
      - description: Filter by observer type (e.g. meshcoretomqtt, meshcore-ha)
        in: query
        name: type
        type: string
      - description: Filter by broker name
        in: query
        name: broker
        type: string
      - description: Filter by status (online or offline)
        in: query
        name: status
        type: string
      - description: Partial case-insensitive display name match
        in: query
        name: name
        type: string
      - description: 'Filter by transport scope name e.g. %23bc (URL-encoded #bc)'
        in: query
        name: scope
        type: string
      - description: last_seen epoch ms of last item for pagination
        in: query
        name: cursor
        type: integer
      - description: Max results (default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_ObserverSummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List observers
      tags:
      - Observers
  /observers/{observerId}:
    get:
      parameters:
      - description: Observer UUID
        in: path
        name: observerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Observer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get observer detail
      tags:
      - Observers
  /observers/{observerId}/adverts:
    get:
      parameters:
      - description: Observer UUID
        in: path
        name: observerId
        required: true
        type: string
      - description: Observation ID of last item for pagination
        in: query
        name: cursor
        type: integer
      - description: Max results (default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_AdvertObservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List advert packets heard by an observer
      tags:
      - Observers
  /observers/{observerId}/telemetry:
    get:
      parameters:
      - description: Observer UUID
        in: path
        name: observerId
        required: true
        type: string
      - description: Duration window e.g. 24h, 48h, 168h (default 24h)
        in: query
        name: range
        type: string
      - description: Return points after this telemetry ID for WS reconnection backfill
        in: query
        name: afterId
        type: integer
      - description: 'Bucketing interval: 1h (default), 6h, or 24h'
        in: query
        name: interval
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetry'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get observer telemetry history
      tags:
      - Observers
...
schemes:
- http
- https
swagger: "2.0"
...